Correctly set configure options
LSI needs to have both 32-bit and 64-bit builds to correctly support
all games, as there is a very large number of 32-bit builds on the
Steam store. To help configure this behaviour, the frontend must be
enabled for the native 64-bit build. However, we're only configuring
the 32-bit for the future right now and not enabling a build of it
until it is ready to be tested.

Additionally we need to inform LSI we're using the new libcxx ABI in
libstdc++, so that additional workarounds are activated to prevent
breakage with C++ display drivers, etc.

This workaround will also ensure games cannot use a vendored version
of libstdc++ which would potentially cause security issues if left
to defaults.

At this point I'm not changing the mask location or shim behaviour
for LSI until we're ready to link Steam/LSI within the repositories.

Signed-off-by: Ikey Doherty <ikey.doherty@intel.com>